Move BridgeTx into SystemTx
