+++ lib/ain-evm/src/core.rs
@@ -7,7 +7,7 @@ use crate::receipt::ReceiptService;
 use crate::services::SERVICES;
 use crate::storage::traits::{BlockStorage, PersistentStateError};
 use crate::storage::Storage;
-use crate::transaction::bridge::{BalanceUpdate, BridgeTx};
+use crate::transaction::system::{BalanceUpdate, SystemTx};
 use crate::trie::TrieDBStore;
 use crate::txqueue::{QueueError, QueueTx, TransactionQueueMap};
 use crate::{
@@ -297,7 +297,7 @@ impl EVMCoreService {
         amount: U256,
         hash: NativeTxHash,
     ) -> Result<(), EVMError> {
-        let queue_tx = QueueTx::BridgeTx(BridgeTx::EvmIn(BalanceUpdate { address, amount }));
+        let queue_tx = QueueTx::SystemTx(SystemTx::EvmIn(BalanceUpdate { address, amount }));
         self.tx_queues
             .queue_tx(queue_id, queue_tx, hash, U256::zero(), U256::zero())?;
         Ok(())
@@ -323,7 +323,7 @@ impl EVMCoreService {
             })
             .into())
         } else {
-            let queue_tx = QueueTx::BridgeTx(BridgeTx::EvmOut(BalanceUpdate { address, amount }));
+            let queue_tx = QueueTx::SystemTx(SystemTx::EvmOut(BalanceUpdate { address, amount }));
             self.tx_queues
                 .queue_tx(queue_id, queue_tx, hash, U256::zero(), U256::zero())?;
             Ok(())
