+++ lib/ain-evm/src/transaction/mod.rs
@@ -1,4 +1,3 @@
-pub mod bridge;
 pub mod system;
 
 use crate::ecrecover::{public_key_to_address, recover_public_key};
