+++ lib/ain-evm/src/transaction/system.rs
@@ -16,7 +16,25 @@ pub struct DST20Data {
 }
 
 #[derive(Debug, Clone)]
+pub struct BalanceUpdate {
+    pub address: H160,
+    pub amount: U256,
+}
+
+#[derive(Debug, Clone)]
 pub enum SystemTx {
     DeployContract(DeployContractData),
     DST20Bridge(DST20Data),
+    EvmIn(BalanceUpdate),
+    EvmOut(BalanceUpdate),
+}
+
+impl SystemTx {
+    pub fn sender(&self) -> Option<H160> {
+        match self {
+            SystemTx::EvmIn(tx) => Some(tx.address),
+            SystemTx::EvmOut(tx) => Some(tx.address),
+            _ => None,
+        }
+    }
 }
